add nav menu that fades-in when scrolling reaches about me
