+++ package.json
@@ -13,6 +13,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@react-hook/window-scroll": "^1.3.0",
     "@types/ramda": "0.30.0",
     "glob": "10.3.15",
     "koa": "2.15.3",
