+++ src/ui/education.ts
@@ -1,0 +1,23 @@
+const educationList = [{
+  logo: '/public/assets/unsw.png',
+  timeframe: '1996 to 2000',
+  school: 'University of New South Wales',
+  certification: 'Bachelor of Electrical Engineering',
+  summary: `<ul>
+              <li>Completed all units required. But did not finish work experience</li>
+              <li>Achieved High Distinction in computer science units</li>
+            </ul>
+            `
+}, {
+  logo: '/public/assets/east_hills_boys_hs.png',
+  timeframe: '1991 to 1996',
+  school: 'East Hills Boys Technology High School',
+  certification: 'High School',
+  summary: `<ul>
+              <li>Awarded Dux for year 12.</li>
+              <li>Presented school in the debating team</li>
+            </ul>`
+}];
+
+export default educationList;
+
