+++ src/ui/skills.ts
@@ -1,0 +1,35 @@
+export const softwareSkills = [
+  'javascript',
+  'nodejs',
+  'react',
+  'kotlin',
+  'server side react',
+  'bun',
+  'swift',
+  'postgresql',
+  'cosmosdb',
+  'mongodb',
+  'graphql',
+  'typescript',
+  'figma',
+  'webpack',
+];
+
+export const devopsSkills = [
+  'azure',
+  'gcp',
+  'aws',
+  'kubernetes',
+  'route53',
+  'terraform',
+  'docker',
+  'linux',
+  'apigee',
+  'bash',
+  'powershell',
+  'jenkins',
+  'sumologic',
+  'splunk',
+  'rancher',
+  'grafana',
+]
