+++ src/ui/styles.css
@@ -135,7 +135,7 @@ body {
 }
 .education-sticky {
   position: sticky;
-  top: 0;
+  top: 48px;
   font-size: 1.8em;
 }
 .education-menu ul {
@@ -188,6 +188,9 @@ body {
 }
 .resume-content h4 {
   color: #219EBC;
+}
+.resume-summary ul {
+  list-style-type: square;
 }
 .about-me {
   padding: 16px;
@@ -248,4 +251,45 @@ body {
   to {
     transform:translate(-50%, -50%)
   }
+}
+
+.navigation-menu {
+  padding: 16px 8px;
+  font-size: 1.3em;
+  position: fixed;
+  top: 0;
+  right: 0;
+  width: 100%;
+  height: 36px;
+  z-index: 4;
+  background-color: #fff;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  align-items: flex-end;
+  justify-content: flex-end;
+}
+.navigation-menu ul {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+.navigation-menu li {
+  float: left;
+  padding: 0 0 0 24px;
+  cursor: pointer;
+}
+.navigation-menu li a {
+  color: #FB8500;
+  text-decoration: none;
+}
+
+.shadow {
+  box-shadow: 0 3px 8px -5px #c1c1c1;
+  animation: fadein 0.2s ease-in
+}
+
+@keyframes fadein {
+  from { opacity: 0; }
+  to   { opacity: 1; }
 }
